test: share the JWKS test server between the two physical-tenant ITs
JwksTestServer and signForIssuer were byte-identical in both ITs. Moving
them to a package-private class in the same package leaves one copy and
exercises both branches of the audiences check, which only one caller
reached before.
